requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.40.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-contrib-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pycocotools==2.0.6
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
